pyproject.toml
[project]
name = "django-basemodels-rest"
version = "0.0.1"
description = "DRF-интеграция для django-basemodels — готовый сериализатор BaseModel и минимальные утилиты для работы с REST."
authors = [
    { name = "skulap", email = "skulap01@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django-basemodels (>=0.0.4,<0.0.5)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "django_basemodels_rest", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)",
    "ruff (>=0.13.0,<0.14.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
]
